Improve CLI argument reading
