+++ openapi-generator/src/main/kotlin/org/jellyfin/openapi/Main.kt
@@ -4,26 +4,12 @@ import org.jellyfin.openapi.hooks.hooksModule
 import org.koin.dsl.koinApplication
 import java.io.File
 
-fun Array<out String>.asArguments(): Map<String, String> {
-	val iterator = iterator()
-	var key: String? = null
-	val arguments = mutableMapOf<String, String>()
-
-	while (iterator.hasNext()) {
-		val value = iterator.next()
-		if (value.startsWith("--")) key = value.removePrefix("--")
-		else if (key != null) arguments[key] = value
-	}
-
-	return arguments
-}
-
 fun main(vararg args: String) {
 	// Read arguments
 	val arguments = args.asArguments()
-	val openApiFile = arguments["openApiFile"] ?: throw Error("Missing argument --openApiFile")
-	val apiOutputDir = arguments["apiOutputDir"] ?: throw Error("Missing argument --apiOutputDir")
-	val modelsOutputDir = arguments["modelsOutputDir"] ?: throw Error("Missing argument --modelsOutputDir")
+	val openApiFile by arguments
+	val apiOutputDir by arguments
+	val modelsOutputDir by arguments
 
 	// Read OpenAPI json
 	val openApiJson = File(openApiFile).readText()
